requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-surprise
wordcloud
jupyter